README.md
SCP: CB Unity Remake Decompiled

SCP: CB Unity Remake Decompiled

This repository contains the decompiled source code for the Unity game "SCP: CB Unity Remake." The original game was developed by Aerie Gaming Studios, and it was intended to be a remake of the popular horror game "SCP: Containment Breach".

Unfortunately, development on the project was officially cancelled in 2021, leaving many fans disappointed. However, I managed to obtain a decompiled version of the game, which I am sharing here for educational and modding purposes.

Please note that this decompiled version may not be complete or fully functional, and it is provided "as is" without any guarantees. However, feel free to explore the code, make modifications, and create your own custom versions of the game. 

Remember to respect the original developers' work and adhere to any relevant copyright or licensing restrictions.

Happy modding!

- JoZeph